requirements.txt
# AI Drug Discovery Modeling - Requirements

# 데이터 처리 및 분석
pandas>=1.3.0
numpy>=1.21.0

# 엑셀 파일 읽기
openpyxl>=3.0.0
xlrd>=2.0.0

# 데이터 시각화 (선택사항)
matplotlib>=3.5.0
seaborn>=0.11.0

# 머신러닝 (선택사항)
scikit-learn>=1.0.0

# RDKit (화학 구조 처리, 선택사항)
rdkit-pypi>=2022.9.1

# Jupyter 노트북 (선택사항)
# jupyter>=1.0.0
# ipykernel>=6.0.0 